fixed compiler warnings
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@27200 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/dvdplayer/DVDCodecs/DVDCodecUtils.cpp
@@ -186,7 +186,7 @@ DVDVideoPicture* CDVDCodecUtils::ConvertToNV12Picture(DVDVideoPicture *pSrc)
       // copy luma
       uint8_t *s = pSrc->data[0];
       uint8_t *d = pPicture->data[0];
-      for (int y = 0; y < pSrc->iHeight; y++)
+      for (int y = 0; y < (int)pSrc->iHeight; y++)
       {
         fast_memcpy(d, s, pSrc->iWidth);
         s += pSrc->iLineSize[0];
@@ -195,11 +195,11 @@ DVDVideoPicture* CDVDCodecUtils::ConvertToNV12Picture(DVDVideoPicture *pSrc)
 
       //copy chroma
       uint8_t *s_u, *s_v, *d_uv;
-      for (int y = 0; y < pSrc->iHeight/2; y++) {
+      for (int y = 0; y < (int)pSrc->iHeight/2; y++) {
         s_u = pSrc->data[1] + (y * pSrc->iLineSize[1]);
         s_v = pSrc->data[2] + (y * pSrc->iLineSize[2]);
         d_uv = pPicture->data[1] + (y * pPicture->iLineSize[1]);
-        for (int x = 0; x < pSrc->iWidth/2; x++) {
+        for (int x = 0; x < (int)pSrc->iWidth/2; x++) {
           *d_uv++ = *s_u++;
           *d_uv++ = *s_v++;
         }
